requirements.txt
pynput==1.7.6
pyperclip==1.8.2
llama-cpp-python==0.2.11
huggingface_hub==0.19.4 